gc.h: Add SET_MACHINE_STACK_END specific for Powerpc64
Currently we are not able to grab the correct end/top of the stack on
powerpc64 due to the fact that it uses the fallback function.

The fallback function does not return the correct top of the stack
because it adds a new frame and the returned top of the stack contains
this frame overhead that could be something around 112 bytes on
Powerpc64.

This patch simply gets the correct top of the stack pointer, which is
always on r1 register according to both ABI v1 and ABI v2 (Little
endian).

Signed-off-by: Breno Leitao <leitao@debian.org>